add: README-socket, remove: all unnessesary files and dependencies
