+++ package.json
@@ -16,13 +16,9 @@
         "@hapi/hapi": "^20.1.0",
         "@hapi/inert": "^6.0.3",
         "fastest-validator": "^1.10.1",
-        "fastify": "^3.14.0",
-        "fastify-cors": "^5.2.0",
         "g": "^2.0.1",
-        "sqlite3": "^5.0.2",
         "ts-loader": "^8.0.17",
         "typescript": "^4.1.5",
-        "uWebSockets.js": "uNetworking/uWebSockets.js#v18.14.0",
         "uuid": "^8.3.2",
         "webpack": "4.44.2",
         "webpack-cli": "3.3.12",
