+++ src/README-socket.md
@@ -1,0 +1,33 @@
+# Online Game with WS
+
+Для того, чтобы сделать фронт-энд рекомендуется:
+ 1. Создать роут Online-Board,
+ 2. Создать компонент для отображения комнат по этому роуту (может быть обычной таблицей),
+ 3. Создать отображения для Player1, Board и Player2,
+ 4. Инициализировать сокет соединение в контексте(скоупе) самого роута,
+
+Обмен данными клиент:
+ - Каждое сообщение должно быть обьектом JSON, сериализованным в строку и содержать поля { type, data }.
+ - type бывает 3-х видов 'createRoom', 'joinRoom', 'playerTurn'.
+ - data отличается от типа, подробнее смотреть файл validation.js, там расписаны все схемы.
+
+Обмен данными сервер: 
+ - Сервер шлет ошибку если что то пошло не так, и в случае критических ошибок закрывает соединение.
+ - Сервер всегда отправляет одинаковый тип сообщения.
+ - rooms: {
+	 	[roomId]: {
+			 userId,
+			 username, // может не быть
+			 roomname, // может не быть
+		 }
+ 	}
+ - game: {
+	 	roomId,
+	 	hands: { p1, p2 } // после createRoom p2 = {} отсюда брать карточки для рисования,
+		board, // состояние доски
+	}
+ - когда игра закочилась отсылает изначальные руки: {
+	 	p1: { pokemons, userId, owner: p1/p2 },
+		p2: { pokemons, userId, owner: p1/p2 }
+ 	}
+
