--- src/game-socket/.gitignore
@@ -1,1 +1,0 @@
-example
