--- src/game-socket/test.js
@@ -1,12 +1,0 @@
-const player = {
-    owner: 'p1',
-    pokemons: [1, 2, 3],
-};
-
-const ros = {
-    p1: [3, 4, 5],
-};
-
-const res = ros[player.owner];
-
-console.log();
