--- src/game-socket/test.json
@@ -1,101 +1,0 @@
-{
-    "move": {
-        "hits": [1, 2, 3, 2],
-        "position": 1,
-        "owner": "p1",
-        "holder": "p1",
-        "rate": 1008
-    },
-    "board": [
-        {
-            "hits": [2, 3, 3, 5],
-            "position": 0,
-            "owner": "p1",
-            "holder": "p1",
-            "rate": 1013
-        },
-        {
-            "hits": [1, 2, 3, 2],
-            "position": 1,
-            "owner": "p1",
-            "holder": "p1",
-            "rate": 1008
-        },
-        {
-            "hits": [2, 2, 3, 4],
-            "position": 2,
-            "owner": "p2",
-            "holder": "p2",
-            "rate": 1011
-        },
-        {
-            "hits": [2, 1, 5, 3],
-            "position": 3,
-            "owner": "p1",
-            "holder": "p1",
-            "rate": 1011
-        },
-        {
-            "hits": [2, 5, 1, 4],
-            "position": 4,
-            "owner": "p1",
-            "holder": "p1",
-            "rate": 1012
-        },
-        {
-            "hits": [3, 1, 4, 2],
-            "position": 5,
-            "owner": "p2",
-            "holder": "p2",
-            "rate": 1010
-        },
-        {
-            "hits": [1, 3, 2, 4],
-            "position": 6,
-            "owner": "p1",
-            "holder": "p1",
-            "rate": 1010
-        },
-        {
-            "hits": [1, 1, 2, 2],
-            "position": 7,
-            "owner": "p2",
-            "holder": "p2",
-            "rate": 1006
-        },
-        {
-            "hits": [1, 2, 2, 1],
-            "position": 8,
-            "owner": "p2",
-            "holder": "p2",
-            "rate": 1006
-        }
-    ],
-    "rate": 0,
-    "beaten": [],
-    "empty": [1],
-    "hands": {
-        "p1": {
-            "owner": "p1",
-            "pokes": [
-                {
-                    "owner": "p1",
-                    "holder": "p1",
-                    "hits": [1, 2, 3, 2],
-                    "rate": 1008
-                }
-            ]
-        },
-        "p2": {
-            "owner": "p2",
-            "pokes": [
-                {
-                    "owner": "p2",
-                    "holder": "p2",
-                    "hits": [10, 7, 10, 9],
-                    "rate": 1036
-                }
-            ]
-        }
-    }
-}
